Initialized App-Routing for a modern single page application. Initializing of an http request to get the xml data of the server later on. Generation of a vaccination service as well as a component for the first page of the document overview page. the data.service was a trial to load an offline database to get some content for testing. Will be deleted soon.
Next Todos: Generatiing basic framework for all Pages of the document
Implementing the template for all seperate pages.
Implementing arrows as well as a menu to navigate through the document.
Setting a fixed ratio of the main yellow box, to generate a scalable view. (Try to do it with javascript not CSS ( no absolute divs)
Load the structure of the information into the view ->Implementing a parser for XML or JSON data.

I know its way to complicated for this simple excercise. But Angular 8 is awesome.
So deal with it or die trying :P